+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 ## [`master`](https://github.com/elastic/eui/tree/master)
 
-No public interface changes since `7.0.0`.
+- Added `text` as a color option for `EuiLink` ([#1571](https://github.com/elastic/eui/pull/1571))
+- Added `EuiResizeObserver` to expose ResizeObserver API to React components; falls back to MutationObserver API in unsupported browsers ([#1559](https://github.com/elastic/eui/pull/1559))
+
+**Bug fixes**
+
+- Fixed content cut off in `EuiContextMenuPanel` when height changes dynamically ([#1559](https://github.com/elastic/eui/pull/1559))
+
+## [`7.1.0`](https://github.com/elastic/eui/tree/v7.1.0)
+
+- Added `append` prop to `EuiFieldText` ([#1567](https://github.com/elastic/eui/pull/1567))
+- Adjusted set of Elastic Logos in `EuiIcon` to look better in dark mode. ([#1462](https://github.com/elastic/eui/pull/1562))
+- Added `isCopyable` prop to `EuiCodeBlock` ([#1556](https://github.com/elastic/eui/pull/1556))
+- Added optional `Snippet` tab to docs and renamed demo tabs ([#1556](https://github.com/elastic/eui/pull/1556))
+- Expanded `getSecureRelForTarget` to handle elastic.co domains as a referrer whitelist ([#1565](https://github.com/elastic/eui/pull/1565))
+- New `url` utility for verifying if a URL is a referrer whitelist ([#1565](https://github.com/elastic/eui/pull/1565))
+- Add iconSize to ButtonIcon type definition ([#1568](https://github.com/elastic/eui/pull/1568))
 
 ## [`7.0.0`](https://github.com/elastic/eui/tree/v7.0.0)
 
@@ -24,7 +39,7 @@ No public interface changes since `7.0.0`.
 
 ## [`6.10.1`](https://github.com/elastic/eui/tree/v6.10.1)
 
-**Note: this release is a backport containing changes originally made in `6.11.0`**
+**Note: this release is a backport containing changes originally made in `7.0.0`**
 
 **Bug fixes**